refactor: extract shared image source resolution util
